refactor(drawingml): extract shared Rust core
