+++ crates/docx-layout/Cargo.toml
@@ -18,6 +18,7 @@ crate-type = ["cdylib", "rlib"]
 wasm-opt = ['-O3']
 
 [dependencies]
+ooxml-drawingml = { path = "../ooxml-drawingml", version = "=0.0.0" }
 ooxml-text = { path = "../ooxml-text" }
 wasm-bindgen = "0.2"
 serde = { version = "1", features = ["derive"] }
