+++ crates/docx-parse/src/drawingml.rs
@@ -6,142 +6,19 @@
 
 use std::collections::HashMap;
 
-use serde::{Deserialize, Serialize};
-
 use crate::scalars::ColorValue;
 use crate::xml::{XmlElement, parse_javascript_integer_prefix};
 
-const ELLIPSE_KAPPA: f64 = 0.552_284_749_830_793_6;
-const ROUND_RECT_RADIUS: f64 = 0.166_666_666_666_666_66;
+pub use ooxml_drawingml::{
+    GeometryPathCommand, GradientFill, GradientStop, LineEnd, Point2D, ShapeFill, ShapeOutline,
+    Size2D, Transform2D, TransformResult, preset_geometry_to_path, resolve_color_value_to_hex,
+};
+
 const ANGLE_UNITS_PER_DEGREE: f64 = 60_000.0;
 const MAX_CUSTOM_PATH_COMMANDS: usize = 2_048;
 const MAX_CUSTOM_GUIDES: usize = 512;
 const MAX_ABS_CUSTOM_PATH_NUMBER: f64 = 1_000_000_000.0;
 const MAX_ABS_NORMALIZED_PATH_NUMBER: f64 = 10_000.0;
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ShapeFill {
-    #[serde(rename = "type")]
-    pub fill_type: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub color: Option<ColorValue>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub gradient: Option<GradientFill>,
-}
-
-impl ShapeFill {
-    fn named(fill_type: &str) -> Self {
-        Self {
-            fill_type: fill_type.to_owned(),
-            color: None,
-            gradient: None,
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub struct GradientFill {
-    #[serde(rename = "type")]
-    pub gradient_type: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub angle: Option<f64>,
-    pub stops: Vec<GradientStop>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub struct GradientStop {
-    pub position: f64,
-    pub color: ColorValue,
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ShapeOutline {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub width: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub color: Option<ColorValue>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub style: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cap: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub join: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub head_end: Option<LineEnd>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tail_end: Option<LineEnd>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub struct LineEnd {
-    #[serde(rename = "type")]
-    pub end_type: String,
-    // Not skipped: getAttribute returns null and the TS object retains it.
-    pub width: Option<String>,
-    pub length: Option<String>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "camelCase")]
-pub enum GeometryPathCommand {
-    Move {
-        x: f64,
-        y: f64,
-    },
-    Line {
-        x: f64,
-        y: f64,
-    },
-    Quad {
-        cpx: f64,
-        cpy: f64,
-        x: f64,
-        y: f64,
-    },
-    Cubic {
-        cp1x: f64,
-        cp1y: f64,
-        cp2x: f64,
-        cp2y: f64,
-        x: f64,
-        y: f64,
-    },
-    Close,
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-pub struct Size2D {
-    pub width: f64,
-    pub height: f64,
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Transform2D {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub rotation: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub flip_h: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub flip_v: Option<bool>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub struct Point2D {
-    pub x: f64,
-    pub y: f64,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub struct TransformResult {
-    pub size: Size2D,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub transform: Option<Transform2D>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub offset: Option<Point2D>,
-}
 
 pub fn parse_color_element(element: Option<&XmlElement>) -> Option<ColorValue> {
     let element = element?;
@@ -270,36 +147,6 @@ pub fn parse_outline(sp_pr: Option<&XmlElement>) -> Option<ShapeOutline> {
     })
 }
 
-pub fn resolve_color_value_to_hex(color: Option<&ColorValue>) -> Option<String> {
-    let color = color?;
-    if let Some(rgb) = &color.rgb {
-        return Some(format!("#{rgb}"));
-    }
-    color
-        .theme_color
-        .as_deref()
-        .map(default_theme_color)
-        .map(|rgb| format!("#{rgb}"))
-}
-
-fn default_theme_color(slot: &str) -> &str {
-    match slot {
-        "dk1" | "text1" => "000000",
-        "lt1" | "background1" => "FFFFFF",
-        "dk2" | "text2" => "44546A",
-        "lt2" | "background2" => "E7E6E6",
-        "accent1" => "4472C4",
-        "accent2" => "ED7D31",
-        "accent3" => "A5A5A5",
-        "accent4" => "FFC000",
-        "accent5" => "5B9BD5",
-        "accent6" => "70AD47",
-        "hlink" => "0563C1",
-        "folHlink" => "954F72",
-        _ => "000000",
-    }
-}
-
 pub fn parse_gradient_fill(element: &XmlElement) -> ShapeFill {
     let mut gradient_type = "linear";
     let mut angle = None;
@@ -412,325 +259,6 @@ pub fn parse_preset_geometry_path(sp_pr: Option<&XmlElement>) -> Option<Vec<Geom
         }
     }
     preset_geometry_to_path(shape_type, &adjustments)
-}
-
-pub fn preset_geometry_to_path(
-    shape_type: &str,
-    adjustments: &HashMap<String, f64>,
-) -> Option<Vec<GeometryPathCommand>> {
-    use GeometryPathCommand as C;
-    let result = match shape_type {
-        "rect" => vec![
-            C::Move { x: 0.0, y: 0.0 },
-            C::Line { x: 1.0, y: 0.0 },
-            C::Line { x: 1.0, y: 1.0 },
-            C::Line { x: 0.0, y: 1.0 },
-            C::Close,
-        ],
-        "roundRect" => {
-            let r = ROUND_RECT_RADIUS;
-            vec![
-                C::Move { x: r, y: 0.0 },
-                C::Line { x: 1.0 - r, y: 0.0 },
-                C::Quad {
-                    cpx: 1.0,
-                    cpy: 0.0,
-                    x: 1.0,
-                    y: r,
-                },
-                C::Line { x: 1.0, y: 1.0 - r },
-                C::Quad {
-                    cpx: 1.0,
-                    cpy: 1.0,
-                    x: 1.0 - r,
-                    y: 1.0,
-                },
-                C::Line { x: r, y: 1.0 },
-                C::Quad {
-                    cpx: 0.0,
-                    cpy: 1.0,
-                    x: 0.0,
-                    y: 1.0 - r,
-                },
-                C::Line { x: 0.0, y: r },
-                C::Quad {
-                    cpx: 0.0,
-                    cpy: 0.0,
-                    x: r,
-                    y: 0.0,
-                },
-                C::Close,
-            ]
-        }
-        "ellipse" => vec![
-            C::Move { x: 1.0, y: 0.5 },
-            C::Cubic {
-                cp1x: 1.0,
-                cp1y: 0.5 + ELLIPSE_KAPPA / 2.0,
-                cp2x: 0.5 + ELLIPSE_KAPPA / 2.0,
-                cp2y: 1.0,
-                x: 0.5,
-                y: 1.0,
-            },
-            C::Cubic {
-                cp1x: 0.5 - ELLIPSE_KAPPA / 2.0,
-                cp1y: 1.0,
-                cp2x: 0.0,
-                cp2y: 0.5 + ELLIPSE_KAPPA / 2.0,
-                x: 0.0,
-                y: 0.5,
-            },
-            C::Cubic {
-                cp1x: 0.0,
-                cp1y: 0.5 - ELLIPSE_KAPPA / 2.0,
-                cp2x: 0.5 - ELLIPSE_KAPPA / 2.0,
-                cp2y: 0.0,
-                x: 0.5,
-                y: 0.0,
-            },
-            C::Cubic {
-                cp1x: 0.5 + ELLIPSE_KAPPA / 2.0,
-                cp1y: 0.0,
-                cp2x: 1.0,
-                cp2y: 0.5 - ELLIPSE_KAPPA / 2.0,
-                x: 1.0,
-                y: 0.5,
-            },
-            C::Close,
-        ],
-        "line" | "straightConnector1" => {
-            vec![C::Move { x: 0.0, y: 0.0 }, C::Line { x: 1.0, y: 1.0 }]
-        }
-        "triangle" | "isosTriangle" => polygon(&[(0.5, 0.0), (1.0, 1.0), (0.0, 1.0)]),
-        "rtTriangle" => polygon(&[(0.0, 0.0), (1.0, 1.0), (0.0, 1.0)]),
-        "diamond" | "flowChartDecision" => {
-            polygon(&[(0.5, 0.0), (1.0, 0.5), (0.5, 1.0), (0.0, 0.5)])
-        }
-        "parallelogram" => {
-            let i = clamp_fraction(adjustments.get("adj").copied(), 0.25);
-            polygon(&[(i, 0.0), (1.0, 0.0), (1.0 - i, 1.0), (0.0, 1.0)])
-        }
-        "trapezoid" => {
-            let i = clamp_fraction(adjustments.get("adj").copied(), 0.2);
-            polygon(&[(i, 0.0), (1.0 - i, 0.0), (1.0, 1.0), (0.0, 1.0)])
-        }
-        "pentagon" | "flowChartOffpageConnector" => regular_polygon(5),
-        "hexagon" => regular_polygon(6),
-        "heptagon" => regular_polygon(7),
-        "octagon" => regular_polygon(8),
-        "decagon" => regular_polygon(10),
-        "dodecagon" => regular_polygon(12),
-        value if value.starts_with("star") => {
-            let points = value[4..].parse::<usize>().ok()?;
-            if !matches!(points, 4 | 5 | 6 | 7 | 8 | 10 | 12 | 16 | 24 | 32) {
-                return None;
-            }
-            star(points)
-        }
-        "bentConnector2" => bent_connector(2, adjustments.get("adj1").copied()),
-        "bentConnector3" => bent_connector(3, adjustments.get("adj1").copied()),
-        "bentConnector4" => bent_connector(4, adjustments.get("adj1").copied()),
-        "bentConnector5" => bent_connector(5, adjustments.get("adj1").copied()),
-        "curvedConnector2" => curved_connector(2),
-        "curvedConnector3" => curved_connector(3),
-        "curvedConnector4" => curved_connector(4),
-        "curvedConnector5" => curved_connector(5),
-        "rightArrow" => arrow("right", adjustments.get("adj2").copied()),
-        "leftArrow" => arrow("left", adjustments.get("adj2").copied()),
-        "upArrow" => arrow("up", adjustments.get("adj2").copied()),
-        "downArrow" => arrow("down", adjustments.get("adj2").copied()),
-        "leftRightArrow" => polygon(&[
-            (0.0, 0.5),
-            (0.25, 0.0),
-            (0.25, 0.25),
-            (0.75, 0.25),
-            (0.75, 0.0),
-            (1.0, 0.5),
-            (0.75, 1.0),
-            (0.75, 0.75),
-            (0.25, 0.75),
-            (0.25, 1.0),
-        ]),
-        "upDownArrow" => polygon(&[
-            (0.5, 0.0),
-            (1.0, 0.25),
-            (0.75, 0.25),
-            (0.75, 0.75),
-            (1.0, 0.75),
-            (0.5, 1.0),
-            (0.0, 0.75),
-            (0.25, 0.75),
-            (0.25, 0.25),
-            (0.0, 0.25),
-        ]),
-        "chevron" => polygon(&[
-            (0.0, 0.0),
-            (0.65, 0.0),
-            (1.0, 0.5),
-            (0.65, 1.0),
-            (0.0, 1.0),
-            (0.35, 0.5),
-        ]),
-        "homePlate" => polygon(&[(0.0, 0.0), (0.75, 0.0), (1.0, 0.5), (0.75, 1.0), (0.0, 1.0)]),
-        "flowChartProcess"
-        | "flowChartAlternateProcess"
-        | "flowChartPredefinedProcess"
-        | "flowChartInternalStorage"
-        | "flowChartPreparation"
-        | "flowChartManualOperation"
-        | "flowChartMagneticTape"
-        | "flowChartMagneticDisk"
-        | "flowChartMagneticDrum"
-        | "flowChartDisplay"
-        | "textBox" => preset_geometry_to_path("rect", adjustments)?,
-        "flowChartConnector" => preset_geometry_to_path("ellipse", adjustments)?,
-        "flowChartInputOutput" | "flowChartManualInput" => {
-            preset_geometry_to_path("parallelogram", adjustments)?
-        }
-        "flowChartTerminator" => preset_geometry_to_path("roundRect", adjustments)?,
-        _ => return None,
-    };
-    Some(result)
-}
-
-fn polygon(points: &[(f64, f64)]) -> Vec<GeometryPathCommand> {
-    let mut commands = points
-        .iter()
-        .enumerate()
-        .map(|(i, &(x, y))| {
-            if i == 0 {
-                GeometryPathCommand::Move { x, y }
-            } else {
-                GeometryPathCommand::Line { x, y }
-            }
-        })
-        .collect::<Vec<_>>();
-    if !points.is_empty() {
-        commands.push(GeometryPathCommand::Close);
-    }
-    commands
-}
-
-fn regular_polygon(sides: usize) -> Vec<GeometryPathCommand> {
-    polygon(
-        &(0..sides)
-            .map(|i| {
-                let a = -std::f64::consts::PI / 2.0
-                    + i as f64 * std::f64::consts::PI * 2.0 / sides as f64;
-                (0.5 + a.cos() * 0.5, 0.5 + a.sin() * 0.5)
-            })
-            .collect::<Vec<_>>(),
-    )
-}
-
-fn star(points: usize) -> Vec<GeometryPathCommand> {
-    polygon(
-        &(0..points * 2)
-            .map(|i| {
-                let a =
-                    -std::f64::consts::PI / 2.0 + i as f64 * std::f64::consts::PI / points as f64;
-                let r = if i % 2 == 0 { 0.5 } else { 0.225 };
-                (0.5 + a.cos() * r, 0.5 + a.sin() * r)
-            })
-            .collect::<Vec<_>>(),
-    )
-}
-
-fn clamp_fraction(value: Option<f64>, fallback: f64) -> f64 {
-    value
-        .filter(|v| v.is_finite())
-        .map(|v| if v > 1.0 { v / 100_000.0 } else { v })
-        .unwrap_or(fallback)
-        .clamp(0.0, 1.0)
-}
-
-fn arrow(direction: &str, adjustment: Option<f64>) -> Vec<GeometryPathCommand> {
-    let h = clamp_fraction(adjustment, 0.5);
-    polygon(&[
-        (0.0, 0.25),
-        (1.0 - h, 0.25),
-        (1.0 - h, 0.0),
-        (1.0, 0.5),
-        (1.0 - h, 1.0),
-        (1.0 - h, 0.75),
-        (0.0, 0.75),
-    ])
-    .into_iter()
-    .map(|c| match c {
-        GeometryPathCommand::Move { x, y } => {
-            let (x, y) = orient(direction, x, y);
-            GeometryPathCommand::Move { x, y }
-        }
-        GeometryPathCommand::Line { x, y } => {
-            let (x, y) = orient(direction, x, y);
-            GeometryPathCommand::Line { x, y }
-        }
-        c => c,
-    })
-    .collect()
-}
-
-fn orient(direction: &str, x: f64, y: f64) -> (f64, f64) {
-    match direction {
-        "left" => (1.0 - x, y),
-        "up" => (y, 1.0 - x),
-        "down" => (y, x),
-        _ => (x, y),
-    }
-}
-
-fn bent_connector(segments: usize, adjustment: Option<f64>) -> Vec<GeometryPathCommand> {
-    let b = clamp_fraction(adjustment, 0.5);
-    if segments <= 2 {
-        return vec![
-            GeometryPathCommand::Move { x: 0.0, y: 0.0 },
-            GeometryPathCommand::Line { x: b, y: 0.0 },
-            GeometryPathCommand::Line { x: b, y: 1.0 },
-            GeometryPathCommand::Line { x: 1.0, y: 1.0 },
-        ];
-    };
-    let mut c = vec![GeometryPathCommand::Move { x: 0.0, y: 0.0 }];
-    for i in 1..segments {
-        let f = i as f64 / segments as f64;
-        let (x, y) = if i % 2 == 1 {
-            (if i == 1 { b } else { f }, (i - 1) as f64 / segments as f64)
-        } else {
-            ((i - 1) as f64 / segments as f64, f)
-        };
-        c.push(GeometryPathCommand::Line { x, y })
-    }
-    c.push(GeometryPathCommand::Line { x: 1.0, y: 1.0 });
-    c
-}
-
-fn curved_connector(segments: usize) -> Vec<GeometryPathCommand> {
-    if segments <= 2 {
-        return vec![
-            GeometryPathCommand::Move { x: 0.0, y: 0.0 },
-            GeometryPathCommand::Cubic {
-                cp1x: 0.5,
-                cp1y: 0.0,
-                cp2x: 0.5,
-                cp2y: 1.0,
-                x: 1.0,
-                y: 1.0,
-            },
-        ];
-    };
-    let mut c = vec![GeometryPathCommand::Move { x: 0.0, y: 0.0 }];
-    for i in 0..segments - 1 {
-        let s = i as f64 / (segments - 1) as f64;
-        let e = (i + 1) as f64 / (segments - 1) as f64;
-        c.push(GeometryPathCommand::Cubic {
-            cp1x: s + (e - s) * 0.5,
-            cp1y: s,
-            cp2x: s + (e - s) * 0.5,
-            cp2y: e,
-            x: e,
-            y: e,
-        })
-    }
-    c
 }
 
 pub fn parse_custom_geometry_path(
