+++ crates/ooxml-drawingml/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "ooxml-drawingml"
+version = "0.0.0"
+edition = "2024"
+license = "Apache-2.0"
+publish = false
+description = "Shared DrawingML colors, themes, shape models, and preset geometry."
+
+[dependencies]
+indexmap = { version = "2", features = ["serde"] }
+serde = { version = "1", features = ["derive"] }
