+++ crates/ooxml-drawingml/src/lib.rs
@@ -1,0 +1,15 @@
+//! Shared DrawingML models and resolution.
+
+mod chart;
+mod color;
+mod geometry;
+mod picture;
+mod shape;
+mod theme;
+
+pub use chart::*;
+pub use color::*;
+pub use geometry::*;
+pub use picture::*;
+pub use shape::*;
+pub use theme::*;
